# public/openapi.yaml
openapi: 3.0.3
info:
  title: University Examination & Result Processing API
  version: 1.0.0
  description: |
    Backend for examinations, marks entry (interactive + bulk CSV), asynchronous
    result processing and publication.

    **Lifecycle**: `draft → marks_entry → marks_locked → processing → results_ready → published`
    (see README for the full state machine).

    **Conventions**
    - Errors use `{"error": {"code", "message", "details?"}}` with stable `code` values.
    - Send `Idempotency-Key` on any POST/PUT to make retries safe: a repeated request
      replays the original response (`Idempotent-Replayed: true`).
    - Mark updates use optimistic concurrency: pass `expected_version` (or `If-Match`).
    - Lists use cursor pagination (`next_cursor` → `?cursor=`).
servers:
  - url: /api/v1
    description: The server this page is opened from
  - url: http://localhost:8080/api/v1
    description: Docker (docker compose up)
  - url: http://127.0.0.1:8000/api/v1
    description: Local (php artisan serve)
security:
  - bearerAuth: []
...
paths:
...
  /result-runs/{run}:
    get:
      tags: [Results]
      summary: Result run status, progress and summary
      parameters: [ { name: run, in: path, required: true, schema: { type: integer } } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RunEnvelope' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token from API_KEYS (disabled when API_KEYS is empty). `X-Api-Key` header also accepted.
...
  schemas:
...
    RunEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            id: { type: integer }
            status: { type: string, enum: [processing, completed, failed] }
            total_chunks: { type: integer }
            students_processed: { type: integer }
            summary: { type: object }
...